Fixed #6: Replaced 'float' with 'double'
Updated code generation and unit tests

+++ src/test/java/net/ninjacat/mk54/FunctionalTest.java
@@ -20,7 +20,7 @@ public class FunctionalTest {
 
         mk54.execute();
 
-        assertThat((double) mk54.getX(), is(closeTo(12.566, 0.01)));
+        assertThat(mk54.getX(), is(closeTo(12.566, 0.01)));
     }
 
     @Test
@@ -30,22 +30,22 @@ public class FunctionalTest {
 
         mk54.execute();
 
-        assertThat((double) mk54.getX(), is(closeTo(1.79176, 0.00001)));
+        assertThat(mk54.getX(), is(closeTo(1.79176, 0.00001)));
     }
 
     @Test
     public void testIndirectJumps() throws Exception {
         final Mk54Wrapper mk54 = compileResource("/functional/indirect_jumps.mk");
-        mk54.setMem(3, 5f);
-        mk54.setMem(4, 7f);
-        mk54.setMem(10, 12f);
+        mk54.setMem(3, 5);
+        mk54.setMem(4, 7);
+        mk54.setMem(10, 12);
 
         mk54.execute();
 
-        assertThat(mk54.getX(), is(18.0f));
-        assertThat(mk54.getMem(3), is(4f));
-        assertThat(mk54.getMem(4), is(8f));
-        assertThat(mk54.getMem(10), is(12f));
+        assertThat(mk54.getX(), is(18.0));
+        assertThat(mk54.getMem(3), is(4.0));
+        assertThat(mk54.getMem(4), is(8.0));
+        assertThat(mk54.getMem(10), is(12.0));
     }
 
     private static Mk54Wrapper compileResource(final String resource) throws Exception {
